+++ container_transform/converter.py
@@ -1,16 +1,12 @@
-import six
-
 from .schema import TransformationTypes, ARG_MAP
 
 from .compose import ComposeTransformer
 from .ecs import ECSTransformer
-from .fig import FigTransformer
 from .systemd import SystemdTransformer
 
 TRANSFORMER_CLASSES = {
     TransformationTypes.COMPOSE.value: ComposeTransformer,
     TransformationTypes.ECS.value: ECSTransformer,
-    TransformationTypes.FIG.value: FigTransformer,
     TransformationTypes.SYSTEMD.value: SystemdTransformer,
 }
 
@@ -73,7 +69,7 @@ class Converter(object):
         :return: A output_type container definition
         """
         output = {}
-        for parameter, options in six.iteritems(ARG_MAP):
+        for parameter, options in ARG_MAP.items():
             output_name = options.get(self.output_type, {}).get('name')
             output_required = options.get(self.output_type, {}).get('required')