+++ container_transform/tests/converter_tests.py
@@ -1,3 +1,4 @@
+import json
 from unittest import TestCase
 
 from container_transform.converter import Converter
@@ -5,32 +6,23 @@ from container_transform.converter import Converter
 
 class ConverterTests(TestCase):
 
-    def test_fig_converter(self):
-
-        filename = './container_transform/tests/fig.yml'
-        conv = Converter(filename, 'fig', 'ecs')
-
-        output = conv.convert()
-
-        self.assertIsInstance(output, str)
-
     def test_ecs_converter(self):
         filename = './container_transform/tests/task.json'
-        conv = Converter(filename, 'ecs', 'fig')
+        conv = Converter(filename, 'ecs', 'compose')
 
-        fig_output = conv.convert()
+        compose_output = conv.convert()
 
-        self.assertIsInstance(fig_output, str)
+        self.assertIsInstance(compose_output, str)
 
         self.assertEqual(0, len(conv.messages))
 
     def test_ecs_converter_just_containers(self):
         filename = './container_transform/tests/containers.json'
-        conv = Converter(filename, 'ecs', 'fig')
+        conv = Converter(filename, 'ecs', 'compose')
 
-        fig_output = conv.convert()
+        compose_output = conv.convert()
 
-        self.assertIsInstance(fig_output, str)
+        self.assertIsInstance(compose_output, str)
 
         self.assertEqual(0, len(conv.messages))
 
@@ -38,9 +30,9 @@ class ConverterTests(TestCase):
         filename = './container_transform/tests/task.json'
         conv = Converter(filename, 'ecs', 'compose')
 
-        fig_output = conv.convert()
+        compose_output = conv.convert()
 
-        self.assertIsInstance(fig_output, str)
+        self.assertIsInstance(compose_output, str)
 
         self.assertEqual(0, len(conv.messages))
 
@@ -51,3 +43,28 @@ class ConverterTests(TestCase):
         output = conv.convert()
 
         self.assertIsInstance(output, str)
+
+    def test_compose_converter_v2_to_ecs(self):
+        self.maxDiff = None
+
+        filename = './container_transform/tests/composev2_extended.yml'
+        output_filename = './container_transform/tests/composev2_extended_output.json'
+        conv = Converter(filename, 'compose', 'ecs')
+
+        output = conv.convert()
+        output_dict = json.loads(output)
+
+        output_want = json.load(open(output_filename, 'r'))
+        self.assertDictEqual(output_dict, output_want)
+
+    def test_compose_converter_v2_systemd(self):
+        self.maxDiff = None
+
+        filename = './container_transform/tests/composev2.yml'
+        output_filename = './container_transform/tests/composev2_output.service'
+        conv = Converter(filename, 'compose', 'systemd')
+
+        output = conv.convert()
+
+        output_want = open(output_filename, 'r').read()
+        self.assertEqual(output, output_want)